refactor(ts-oss): extract expiration helpers into a util module
Move EXPIRATION_DATE_PATTERN, normalizeExpirationDate, payloadIsExpired
and todayUtc out of memory/index.ts into utils/expiration.ts. todayUtc
stays module-private since only payloadIsExpired uses it.

Also use a real UUID for the memory id in the Memory-class compat test
instead of "id-1", matching the ids the SDK actually returns.
